requirements.txt
python-telegram-bot==20.7
asyncio-mqtt==0.16.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
asyncpg==0.29.0
aiohttp==3.9.1
